% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeData_mosaic.r
\name{makeData_mosaic}
\alias{makeData_mosaic}
\title{Confusion table to a mosaic plot in ggplot2
makeData_mosaic function generate data for a mosaic plot in ggplot2}
\usage{
makeData_mosaic(data, x, y, ...)
}
\arguments{
\item{data}{a data.frame}

\item{x}{the name of the first variable (a quality factor)}

\item{y}{the name of the second variable (a quality factor)}
}
\value{
a list with informations for ggplot2
}
\description{
Confusion table to a mosaic plot in ggplot2
makeData_mosaic function generate data for a mosaic plot in ggplot2
}

